Enabled to execute migration in mysql / mariadb environment.
Enabled to execute migration in mysql / mariadb environment.
refs #1

The following process fails for mysql / mariadb due to foreign-key constraint.
remove_index :taggings, :tag_id

Even if the index is not deleted, only the useless index remains.
There are other indexes and not dropping them is not a major performance issue.

Added a process to check the index creation status before deleting / adding an index in case migration fails in the middle.

+++ db/migrate/20150326093104_add_missing_unique_indice.rb
@@ -1,8 +1,12 @@
 class AddMissingUniqueIndice < ActiveRecord::Migration[4.2]
   def self.up
-    add_index :tags, :name, unique: true
-    remove_index :taggings, :tag_id
-    remove_index :taggings, [:taggable_id, :taggable_type, :context]
+    unless index_exists? :tags, :name, unique: true
+      add_index :tags, :name, unique: true
+    end
+    # remove_index :taggings, :tag_id
+    if index_exists? :taggings, [:taggable_id, :taggable_type, :context]
+      remove_index :taggings, [:taggable_id, :taggable_type, :context]
+    end
     add_index :taggings,
       [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type],
       unique: true, name: 'taggings_idx'
